Use uint64_t for ticks.
This fixes timer overflow and undefined behavior.
It was observed on long running server that
each 2^31 ticks since usrsctp initialization integer
overflow happen.
2^31 ticks with default time frequency 1000 Hz
(1 tick = 1 ms) overflowed after ~24.8 days, causing
server misbehavior.
With uint64_t used for ticks and default frequency
of 1000 Hz time can overflow only after ~584 billion years.

+++ usrsctplib/netinet/sctp_bsd_addr.c
@@ -948,7 +948,7 @@ sctp_packet_log(struct mbuf *m)
 	lenat++;
 	tick_tock = (uint32_t *)lenat;
 	lenat++;
-	*tick_tock = sctp_get_tick_count();
+	*tick_tock = (uint32_t)sctp_get_tick_count();
 	copyto = (void *)lenat;
 	thisone = thisend - sizeof(int);
 	lenat = (int *)&SCTP_BASE_VAR(packet_log_buffer)[thisone];
